@@ -1,12 +1,12 @@
 [package]
 name = "prover"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 publish = false
 
 [dependencies]
 bincode = "1.3.3"
-brotli2 = "0.3.2"
+derivative = "2.2.0"
 digest = "0.9.0"
 eyre = "0.6.5"
 fnv = "1.0.7"
@@ -15,24 +15,38 @@ libc = "0.2.108"
 nom = "7.0.0"
 nom-leb128 = "0.2.0"
 num = "0.4"
-rayon = "1.5.1"
 rustc-demangle = "0.1.21"
 serde = { version = "1.0.130", features = ["derive", "rc"] }
 serde_json = "1.0.67"
 sha3 = "0.9.1"
 static_assertions = "1.1.0"
 structopt = "0.3.23"
-wasmparser = "0.84.0"
-wat = "1.0.56"
 serde_with = "1.12.1"
-lazy_static = "1.4.0"
+parking_lot = "0.12.1"
+lazy_static.workspace = true
+itertools = "0.10.5"
+wat = "1.0.56"
 smallvec = { version = "1.10.0", features = ["serde"] }
+rayon = { version = "1.5.1", optional = true }
 arbutil = { path = "../arbutil/" }
-c-kzg = "0.4.0" # TODO: look into switching to rust-kzg (no crates.io release or hosted rustdoc yet)
+brotli = { path = "../brotli/" }
+wasmer = { path = "../tools/wasmer/lib/api", optional = true }
+wasmer-types = { path = "../tools/wasmer/lib/types" }
+wasmer-compiler-singlepass = { path = "../tools/wasmer/lib/compiler-singlepass", optional = true, default-features = false, features = ["std", "unwind", "avx"] }
+wasmparser.workspace = true
+num-derive = "0.4.1"
+num-traits = "0.2.17"
+c-kzg = { version = "0.4.0", optional = true } # TODO: look into switching to rust-kzg (no crates.io release or hosted rustdoc yet)
 sha2 = "0.9.9"
 lru = "0.12.3"
 once_cell = "1.19.0"
 
 [lib]
 name = "prover"
-crate-type = ["staticlib","lib"]
+crate-type = ["staticlib", "lib"]
+
+[features]
+default = ["native", "rayon", "singlepass_rayon"]
+native = ["dep:wasmer", "dep:wasmer-compiler-singlepass", "brotli/wasmer_traits", "dep:c-kzg"]
+singlepass_rayon = ["wasmer-compiler-singlepass?/rayon"]
+rayon = ["dep:rayon"]
