@@ -1,21 +1,20 @@
-// Copyright 2022, Offchain Labs, Inc.
+// Copyright 2022-2024, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
 use crate::machine::{Escape, WasmEnv};
-
 use arbutil::{color, Color};
-use structopt::StructOpt;
-use wasmer::Value;
-
+use eyre::Result;
 use std::path::PathBuf;
+use structopt::StructOpt;
 
 mod arbcompress;
-mod gostack;
+mod caller_env;
 mod machine;
-mod runtime;
+mod program;
 mod socket;
-mod syscall;
+mod stylus_backend;
 mod test;
+mod wasip1_stub;
 mod wavmio;
 
 #[derive(StructOpt)]
@@ -45,37 +44,24 @@ pub struct Opts {
     forks: bool,
     #[structopt(long)]
     debug: bool,
+    #[structopt(long)]
+    require_success: bool,
 }
 
-fn main() {
+fn main() -> Result<()> {
     let opts = Opts::from_args();
-
     let env = match WasmEnv::cli(&opts) {
         Ok(env) => env,
-        Err(err) => panic!("{}", err),
+        Err(err) => panic!("{err}"),
     };
 
     let (instance, env, mut store) = machine::create(&opts, env);
 
-    let memory = instance.exports.get_memory("mem").unwrap();
-    let memory = memory.view(&store);
-
-    // To pass in the program name argument, we need to put it in memory.
-    // The Go linker guarantees a section of memory starting at byte 4096 is available for this purpose.
-    // https://github.com/golang/go/blob/252324e879e32f948d885f787decf8af06f82be9/misc/wasm/wasm_exec.js#L520
-    let free_memory_base: i32 = 4096;
-    let name = free_memory_base;
-    let argv = name + 8;
-
-    memory.write(name as u64, b"js\0").unwrap(); // write "js\0" to the name ptr
-    memory.write(argv as u64, &name.to_le_bytes()).unwrap(); // write the name ptr to the argv ptr
-    let run_args = &[Value::I32(1), Value::I32(argv)]; // pass argv with our single name arg
-
-    let main = instance.exports.get_function("run").unwrap();
-    let outcome = main.call(&mut store, run_args);
+    let main = instance.exports.get_function("_start").unwrap();
+    let outcome = main.call(&mut store, &[]);
     let escape = match outcome {
         Ok(outcome) => {
-            println!("Go returned values {:?}", outcome);
+            println!("Go returned values {outcome:?}");
             None
         }
         Err(outcome) => {
@@ -92,6 +78,13 @@ fn main() {
         }
     };
 
+    let memory_used = instance
+        .exports
+        .get_memory("memory")
+        .unwrap()
+        .view(&store)
+        .size();
+
     let env = env.as_mut(&mut store);
     let user = env.process.socket.is_none();
     let time = format!("{}ms", env.process.timestamp.elapsed().as_millis());
@@ -102,11 +95,12 @@ fn main() {
         Some(Escape::Exit(x)) => (false, format!("Failed in {time} with exit code {x}.")),
         Some(Escape::Failure(err)) => (false, format!("Jit failed with {err} in {time}.")),
         Some(Escape::HostIO(err)) => (false, format!("Hostio failed with {err} in {time}.")),
+        Some(Escape::Child(err)) => (false, format!("Child failed with {err} in {time}.")),
         Some(Escape::SocketError(err)) => (false, format!("Socket failed with {err} in {time}.")),
         None => (false, "Machine exited prematurely".to_owned()),
     };
 
-    if opts.debug {
+    if opts.debug || !success {
         println!("{message}");
     }
 
@@ -114,9 +108,13 @@ fn main() {
         true => None,
         false => Some(message),
     };
-    let memory_used = memory.size().0 as u64 * 65_536;
 
     env.send_results(error, memory_used);
+
+    if !success && opts.require_success {
+        std::process::exit(1);
+    }
+    Ok(())
 }
 
 // require a usize be at least 32 bits wide
