@@ -1,25 +1,35 @@
-// Copyright 2021-2023, Offchain Labs, Inc.
+// Copyright 2021-2024, Offchain Labs, Inc.
 // For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
 
 #![allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
 
 pub mod binary;
 mod host;
+#[cfg(feature = "native")]
 mod kzg;
 pub mod machine;
 /// cbindgen:ignore
 mod memory;
 mod merkle;
+mod print;
+pub mod programs;
 mod reinterpret;
 pub mod utils;
 pub mod value;
 pub mod wavm;
 
-use crate::machine::{argument_data_to_inbox, Machine};
-use arbutil::PreimageType;
-use eyre::Result;
+#[cfg(test)]
+mod test;
+
+pub use machine::Machine;
+
+use arbutil::{Bytes32, PreimageType};
+use eyre::{Report, Result};
 use lru::LruCache;
-use machine::{get_empty_preimage_resolver, GlobalState, MachineStatus, PreimageResolver};
+use machine::{
+    argument_data_to_inbox, get_empty_preimage_resolver, GlobalState, MachineStatus,
+    PreimageResolver,
+};
 use once_cell::sync::OnceCell;
 use static_assertions::const_assert_eq;
 use std::{
@@ -27,12 +37,13 @@ use std::{
     num::NonZeroUsize,
     os::raw::{c_char, c_int},
     path::Path,
+    ptr, slice,
     sync::{
         atomic::{self, AtomicU8},
         Arc, Mutex,
     },
 };
-use utils::{Bytes32, CBytes};
+use utils::CBytes;
 
 lazy_static::lazy_static! {
     static ref BLOBHASH_PREIMAGE_CACHE: Mutex<LruCache<Bytes32, Arc<OnceCell<CBytes>>>> = Mutex::new(LruCache::new(NonZeroUsize::new(12).unwrap()));
@@ -58,12 +69,15 @@ pub unsafe extern "C" fn arbitrator_load_machine(
     binary_path: *const c_char,
     library_paths: *const *const c_char,
     library_paths_size: isize,
+    debug_chain: usize,
 ) -> *mut Machine {
-    match arbitrator_load_machine_impl(binary_path, library_paths, library_paths_size) {
+    let debug_chain = debug_chain != 0;
+    match arbitrator_load_machine_impl(binary_path, library_paths, library_paths_size, debug_chain)
+    {
         Ok(mach) => mach,
         Err(err) => {
-            eprintln!("Error loading binary: {}", err);
-            std::ptr::null_mut()
+            eprintln!("Error loading binary: {:?}", err);
+            ptr::null_mut()
         }
     }
 }
@@ -72,6 +86,7 @@ unsafe fn arbitrator_load_machine_impl(
     binary_path: *const c_char,
     library_paths: *const *const c_char,
     library_paths_size: isize,
+    debug_chain: bool,
 ) -> Result<*mut Machine> {
     let binary_path = cstr_to_string(binary_path);
     let binary_path = Path::new(&binary_path);
@@ -88,6 +103,8 @@ unsafe fn arbitrator_load_machine_impl(
         true,
         false,
         false,
+        debug_chain,
+        debug_chain,
         Default::default(),
         Default::default(),
         get_empty_preimage_resolver(),
@@ -96,20 +113,38 @@ unsafe fn arbitrator_load_machine_impl(
 }
 
 #[no_mangle]
+#[cfg(feature = "native")]
 pub unsafe extern "C" fn arbitrator_load_wavm_binary(binary_path: *const c_char) -> *mut Machine {
     let binary_path = cstr_to_string(binary_path);
     let binary_path = Path::new(&binary_path);
     match Machine::new_from_wavm(binary_path) {
         Ok(mach) => Box::into_raw(Box::new(mach)),
         Err(err) => {
-            eprintln!("Error loading binary: {}", err);
-            std::ptr::null_mut()
+            eprintln!("Error loading binary: {err}");
+            ptr::null_mut()
         }
     }
 }
 
 unsafe fn cstr_to_string(c_str: *const c_char) -> String {
     CStr::from_ptr(c_str).to_string_lossy().into_owned()
+}
+
+pub fn err_to_c_string(err: Report) -> *mut libc::c_char {
+    str_to_c_string(&format!("{err:?}"))
+}
+
+/// Copies the str-data into a libc free-able C string
+pub fn str_to_c_string(text: &str) -> *mut libc::c_char {
+    unsafe {
+        let buf = libc::malloc(text.len() + 1); // includes null-terminating byte
+        if buf.is_null() {
+            panic!("Failed to allocate memory for error string");
+        }
+        ptr::copy_nonoverlapping(text.as_ptr(), buf as *mut u8, text.len());
+        *(buf as *mut u8).add(text.len()) = 0;
+        buf as *mut libc::c_char
+    }
 }
 
 #[no_mangle]
@@ -129,22 +164,10 @@ pub unsafe extern "C" fn atomic_u8_store(ptr: *mut u8, contents: u8) {
     (*(ptr as *mut AtomicU8)).store(contents, atomic::Ordering::Relaxed);
 }
 
-fn err_to_c_string(err: eyre::Report) -> *mut libc::c_char {
-    let err = format!("{:#}", err);
-    unsafe {
-        let buf = libc::malloc(err.len() + 1);
-        if buf.is_null() {
-            panic!("Failed to allocate memory for error string");
-        }
-        std::ptr::copy_nonoverlapping(err.as_ptr(), buf as *mut u8, err.len());
-        *(buf.add(err.len()) as *mut u8) = 0;
-        buf as *mut libc::c_char
-    }
-}
-
 /// Runs the machine while the condition variable is zero. May return early if num_steps is hit.
 /// Returns a c string error (freeable with libc's free) on error, or nullptr on success.
 #[no_mangle]
+#[cfg(feature = "native")]
 pub unsafe extern "C" fn arbitrator_step(
     mach: *mut Machine,
     num_steps: u64,
@@ -164,7 +187,7 @@ pub unsafe extern "C" fn arbitrator_step(
         }
         remaining_steps -= stepping;
     }
-    std::ptr::null_mut()
+    ptr::null_mut()
 }
 
 #[no_mangle]
@@ -176,7 +199,7 @@ pub unsafe extern "C" fn arbitrator_add_inbox_message(
 ) -> c_int {
     let mach = &mut *mach;
     if let Some(identifier) = argument_data_to_inbox(inbox_identifier) {
-        let slice = std::slice::from_raw_parts(data.ptr, data.len);
+        let slice = slice::from_raw_parts(data.ptr, data.len);
         let data = slice.to_vec();
         mach.add_inbox_msg(identifier, index, data);
         0
@@ -185,9 +208,22 @@ pub unsafe extern "C" fn arbitrator_add_inbox_message(
     }
 }
 
+/// Adds a user program to the machine's known set of wasms.
+#[no_mangle]
+pub unsafe extern "C" fn arbitrator_add_user_wasm(
+    mach: *mut Machine,
+    module: *const u8,
+    module_len: usize,
+    module_hash: *const Bytes32,
+) {
+    let module = slice::from_raw_parts(module, module_len);
+    (*mach).add_stylus_module(*module_hash, module.to_owned());
+}
+
 /// Like arbitrator_step, but stops early if it hits a host io operation.
 /// Returns a c string error (freeable with libc's free) on error, or nullptr on success.
 #[no_mangle]
+#[cfg(feature = "native")]
 pub unsafe extern "C" fn arbitrator_step_until_host_io(
     mach: *mut Machine,
     condition: *const u8,
@@ -197,10 +233,10 @@ pub unsafe extern "C" fn arbitrator_step_until_host_io(
     while condition.load(atomic::Ordering::Relaxed) == 0 {
         for _ in 0..1_000_000 {
             if mach.is_halted() {
-                return std::ptr::null_mut();
+                return ptr::null_mut();
             }
             if mach.next_instruction_is_host_io() {
-                return std::ptr::null_mut();
+                return ptr::null_mut();
             }
             match mach.step_n(1) {
                 Ok(()) => {}
@@ -208,7 +244,7 @@ pub unsafe extern "C" fn arbitrator_step_until_host_io(
             }
         }
     }
-    std::ptr::null_mut()
+    ptr::null_mut()
 }
 
 #[no_mangle]
@@ -219,7 +255,7 @@ pub unsafe extern "C" fn arbitrator_serialize_state(
     let mach = &*mach;
     let res = CStr::from_ptr(path)
         .to_str()
-        .map_err(eyre::Report::from)
+        .map_err(Report::from)
         .and_then(|path| mach.serialize_state(path));
     if let Err(err) = res {
         eprintln!("Failed to serialize machine state: {}", err);
@@ -237,7 +273,7 @@ pub unsafe extern "C" fn arbitrator_deserialize_and_replace_state(
     let mach = &mut *mach;
     let res = CStr::from_ptr(path)
         .to_str()
-        .map_err(eyre::Report::from)
+        .map_err(Report::from)
         .and_then(|path| mach.deserialize_and_replace_state(path));
     if let Err(err) = res {
         eprintln!("Failed to deserialize machine state: {}", err);
@@ -297,6 +333,7 @@ pub struct ResolvedPreimage {
     pub len: isize, // negative if not found
 }
 
+#[cfg(feature = "native")]
 unsafe fn handle_preimage_resolution(
     context: u64,
     ty: PreimageType,
@@ -325,6 +362,7 @@ unsafe fn handle_preimage_resolution(
 }
 
 #[no_mangle]
+#[cfg(feature = "native")]
 pub unsafe extern "C" fn arbitrator_set_preimage_resolver(
     mach: *mut Machine,
     resolver: unsafe extern "C" fn(u64, u8, *const u8) -> ResolvedPreimage,
@@ -354,16 +392,17 @@ pub unsafe extern "C" fn arbitrator_set_context(mach: *mut Machine, context: u64
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn arbitrator_hash(mach: *mut Machine) -> utils::Bytes32 {
+pub unsafe extern "C" fn arbitrator_hash(mach: *mut Machine) -> Bytes32 {
     (*mach).hash()
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn arbitrator_module_root(mach: *mut Machine) -> utils::Bytes32 {
+pub unsafe extern "C" fn arbitrator_module_root(mach: *mut Machine) -> Bytes32 {
     (*mach).get_modules_root()
 }
 
 #[no_mangle]
+#[cfg(feature = "native")]
 pub unsafe extern "C" fn arbitrator_gen_proof(mach: *mut Machine) -> RustByteArray {
     let mut proof = (*mach).serialize_proof();
     let ret = RustByteArray {
